Added small houses to the terrain generation.
This is a 1x1 size chunk with a brick walled house with a wood roof. It has a simple fire place with two torches to act as a fire. It also spawns a bookcase, 4 chairs, and a table as furniture, but placement is messy, presumably due to the physics of the models. There is also a 50% chance of a treasure chest.