Assorted improvements and bugfixes for lighting.
bugfix: The shader function 'los_blinn_phong' no longer returns negative diffuse brightnesses. This resulted in surfaces facing away from the source being darkened or develop strange colours.

bugfix: Objects with a light effect weren't having their light effect properly attached, this resulted in orphan lights that persist if the source was put away.

bugfix: Local lights were not being applied to the terrain. This is caused by the model of a chunk of terrain being very big. This results in the model appearing to be beyond the maximum range of a light, even though it's just merely huge. THe current fix is to increase the maximum light range to an absurd number. A proper fix will involve either subdividing chunks vertically, or some how changing Ogre's light distance culling strategy.

Gave the torch's light an orangey yellow tint, and changed the attenuation equation for it.