Fix light and particle effect location on carried torches.
bugfix: LIRenObject::find_node() was not searching for nodes if the skeleton wasn't created (which happens when the object isn't animated). For carried items, this would result in the node not being found and the partical effects and/or light being anchored to the carry point. find_node() will now search the model's nodes and use the at rest position if there is no skeleton.

Added limdl_node_get_full_rest_transform() to the C class LIMdlNode. This returns the full transform of a node in the rest position.

bugfix: The Lua class ModelEffect would forget its previously stored special effects each time add_special_effects() is called.