bugfix: Images loaded asynchronously should no longer at risk of being instantly garbage collected.
After an image was asynchronously loaded by the ImageManager's update() function, it would get stored in the table images_by_name, so that waiting coroutines can pick it up later. Unfortunately images_by_name is table of weak references, meaning images stored there can get garbage collected if there are no other references to it. Unfortunately for images loaded like this, this is the only reference, meaning it's a race between the garbage collector and the coroutine retrieving the image. The fix simply uses the seconds_by_image table to give the image a minimum lifespan.