bugfix: Wasn't compiling with Ogre 1.9 .
Ogre 1.9 changed TexturePtr from a subclass of SharedPtr<Texture> to a typedef alias, which causes the compiler to choke when trying to use placement destructors. This bugfix instead uses new placement on _LIExtTexture, so the TexturePtr member is indirectly constructed and destructed.

bugfix: Texture_set_image() now loads new image through the TexturePtr handle, and not through the texture manager.