Added experimental support for VR (outputting to head mounted displays).
Modified the Lua WindowManager class to store and pass around the width and height of the actual viewport, instead of the raw screen/window size.
Modified various UI elements to use Ui.size instead of Graphics:get_video_mode(), as Ui.size stores the usable size of the viewport, instead of the raw screen/window size.

Added a Viewport module and Lua extension for adding and removing view ports. Note: This was a quick and dirty way to add support, and its functionality should probably be transferred to the Camera extension.

bugfix: The outline compositor no longer uses pooled texture, as they were getting used in only 1 view port if more than one view port was in use.