bugfix: The player and other critters should no longer teleport into the void if they're moving slowly.
private_calculate_control() in 'src/lipsofsuna/extension/physics/physics-control-character.cpp' was calculating acceleration incorrectly. This calculation was meant to be a ramp that decreased as the speed got closer to the target, but always resulted in 1, except in certain conditions where it produced a NaN. If a NaN occured this would propogate into a NaN speed and in turn a NaN position. When this happens the affected character's object position will be teleported to (0, 0, 0) (the physics position will remain (NaN, NaN, NaN)).