Added extra source code documentation.
Added a 'main page' to give a broad overview of the codebase.

Added a page on Lua extensions to give a primer on how to write a Lua extension for Lips of Suna.

Improved documentation of C objects by marking member functions with \memberof in sourc documentation. C classes done: _LIMaiProgram, _LIPthPaths